# Conflicts:
#	nts.uk/uk.at/at.web/nts.uk.at.web/src/main/webapp/view/kdw/001/c/kdw001.c.main.css
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/nts.uk.com.web.nittsu.bundles.js
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/nts.uk.com.web.nittsu.bundles.js.map
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/ui/action.ts
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/ui/ko-ext/editor-ko-ext.ts
#	nts.uk/uk.base/base.web/nts.uk.com.js.web/src/main/webapp/lib/nittsu/ui/mgrid.ts
#	nts.uk/uk.com/com.web/nts.uk.com.web/src/main/webapp/view/common/error/common.error.style.css
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/com/mail/MailSender.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/infra/file/report/masterlist/generator/AsposeMasterListGenerator.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/infra/mail/DefaultMailSender.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/infra/web/component/ViewContext.java
#	nts.uk/uk.com/shr/nts.uk.shr.com/src/main/java/nts/uk/shr/infra/web/request/ProgramIdDetector.java
#	nts.uk/uk.repo/nts.arc.jar
#	nts.uk/uk.repo/nts.gul.jar

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/workrecord/erroralarm/condition/attendanceitem/KrcstErAlCompareRange.java
@@ -40,11 +40,11 @@ public class KrcstErAlCompareRange extends UkJpaEntity implements Serializable {
 	@Basic(optional = false)
 	@NotNull
 	@Column(name = "START_VALUE")
-	public int startValue;
+	public double startValue;
 	@Basic(optional = false)
 	@NotNull
 	@Column(name = "END_VALUE")
-	public int endValue;
+	public double endValue;
 	
 	@OneToOne
 	@JoinColumns({
@@ -58,7 +58,7 @@ public class KrcstErAlCompareRange extends UkJpaEntity implements Serializable {
 	}
 
 	public KrcstErAlCompareRange(KrcstErAlCompareRangePK krcstEralCompareRangePK, int compareAtr,
-			int startValue, int endValue) {
+			double startValue, double endValue) {
 		super();
 		this.krcstEralCompareRangePK = krcstEralCompareRangePK;
 		this.compareAtr = compareAtr;
